separate tz_convert tests
